refactor: extract pin/poll overlay key handlers into handlers/keys.rs (closes #417)
Moves execute_pin_toggle, handle_pin_duration_key, and handle_poll_vote_key
out of impl App into a new src/handlers/keys.rs module. These three are the
last user-action handlers in app.rs that needed to reach into handlers/signal.rs
for handle_system_message and handle_poll_vote, forcing those targets to be
pub(crate). With the call sites moved to a sibling module under handlers/,
handle_system_message and handle_poll_vote are now pub(super) instead -- visible
to other handler modules but private from app.rs and the rest of the crate.

Net effect: the bidirectional dependency between app.rs and handlers::signal
is gone. handle_signal_event is the only handlers::signal entry point exposed
to the rest of the crate. app.rs still keeps thin shim methods on App for the
two pub key handlers so existing main-loop call sites need no change.

- src/app.rs: -188 lines, +6 lines (two thin shims + 3 call-site rewrites)
- src/handlers/keys.rs: new 195-line module
- src/handlers/signal.rs: pub(crate) -> pub(super) on two functions
- src/handlers/mod.rs: register new submodule
- 548 tests still passing, clippy clean, fmt clean
- App field-count baseline unchanged (66/66)

Refs CLAUDE.md app field-count ratchet (issue #352): unblocks further App
field extractions by removing call-site coupling that previously alternated
borrows on the App struct.
